+++ src/pages/data-tables.py
@@ -2,7 +2,7 @@ import streamlit as st
 from components.init import init
 from components.menu import menu
 from lib.prefixes import explicits_uri
-from lib.sparql_queries import get_ontology, get_class_tables, get_class_number
+from lib.sparql_queries import get_ontology, get_data_table_columns, get_data_table, get_class_number
 import lib.state as state
 
 
@@ -32,14 +32,14 @@ else:
     st.title("Data tables")
     st.text('')
 
-    col1, col2 = st.columns([5, 1])
+    col1, col2, _, col3, _, col4, col5 = st.columns([2, 1, 1, 2, 1, 2, 2])
 
     # Class filter
     classes_labels = list(map(lambda cls: cls.display_label , ontology.classes))
     class_label = col1.selectbox('Get entity table of class:', options=classes_labels, index=None)
 
     # Number fetched by request
-    limit = col2.selectbox('Limit fetched by request', [10, 20, 50, 100], index=0)
+    limit = col2.selectbox('Limit', [10, 20, 50, 100], index=0)
 
     # Find the selected entity from the selected label
     if class_label:
@@ -49,12 +49,29 @@ else:
         # Offset calculation
         offset = (current_page - 1) * limit
 
+        # Sort by
+        sort_col, sort_way = None, None
+        sort_options = list(map(lambda col: (f"{col}: ASC", f"{col}: DESC"), list(get_data_table_columns().values())))
+        sort_options = [x for tpl in sort_options for x in tpl]
+        sort_option = col3.selectbox('Column to sort on', sort_options, index=None)
+        if sort_option: 
+            sort_col = sort_option[0:sort_option.rindex(':')].strip()
+            sort_way = sort_option[sort_option.rindex(':') + 1:].strip()
+
+
+        # Filter by
+        filter_col, filter_value = None, None
+        filter_col = col4.selectbox('Column to filter by', list(get_data_table_columns().values()), index=None)
+        if filter_col:
+            filter_value = col5.text_input('Value to filter by')
+        
         # Fetch and display data
         st.text('')
-        df_instances = get_class_tables(graph, selected_class_uri, limit, offset)
+        df_instances = get_data_table(graph, selected_class_uri, limit, offset, sort_col=sort_col, sort_way=sort_way, filter_col=filter_col, filter_value=filter_value)
         df_instances.index += 1 * (limit * (current_page - 1)) + 1 # So that indexes appears to start at 1
         if len(df_instances):
             df_instances['Link'] = [f"/?page=entity&endpoint={explicits_uri(endpoint.url)}&graph={explicits_uri(graph.uri)}&entity={explicits_uri(uri)}" for uri in df_instances['URI']]
+        
         st.dataframe(df_instances, use_container_width=True, column_config={'Link':st.column_config.LinkColumn(display_text="Open", width='small')})
 
         # Pagination